extra/requirements.txt
# Core Dependencies
numpy>=1.23
opencv-python>=4.7
pyyaml>=6.0
matplotlib>=3.5
scipy>=1.8
tqdm>=4.64

# Optional but recommended:
# YOLOv8 detection (requires PyTorch, install torch separately depending on your GPU/CPU)
ultralytics>=8.0.0

# Kalman Filtering (optional but improves performance)
filterpy>=1.4.5

# Jupyter notebooks (optional for calibration and testing)
notebook
ipykernel

# Intel RealSense camera support
pyrealsense2>=2.55

# If using PyTorch CPU-only, install with:
# pip install torch==2.0.1+cpu -f https://download.pytorch.org/whl/torch_stable.html

# For GPU builds, install torch following instructions at:
# https://pytorch.org/get-started/locally/
